+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Features:
     launch web servers
   - added support for the PBS Professional resource manager
   - removed deprecated tests for the Slurm adapter
-  - added method to filter list of batch jobs for a given owner
+  - added method to filter list of batch jobs for a given owner or owners
 
 Bugfixes: